+++ CHANGELOG.md
@@ -5,10 +5,13 @@ Changelog for NeoFS Node
 
 ### Added
 ### Changed
+- `object lock` command reads CID and OID the same way other commands do (#1971)
+
 ### Fixed
 ### Removed
 ### Updated
 ### Updating from v0.34.0
+Pass CID and OID parameters via the `--cid` and `--oid` flags, not as the command arguments.
 
 ## [0.34.0] - 2022-10-31 - Marado (마라도, 馬羅島)